Ported ScriptProcessor from Groovy to Java. Fixed bug in ConfigValidator.
